[Fix/#53] 버그 수정 및 에러 메시지 변경
