+++ src/components/room/StudyCalendar.tsx
@@ -57,7 +57,7 @@ function StudyCalendar() {
         value={calendarValue}
         next2Label={null}
         prev2Label={null}
-        formatDay={(date) => moment(date).format('DD')}
+        formatDay={(_, date) => moment(date).format('DD')}
         tileContent={renderTileContent}
         onClickDay={handleDateClick}
       />
